Initial commit
This is the Living As One NASA mars rover exercise initial check-in. This is a SpringBoot application which uses a REST template to walk through the provided date file, dynamically build the GET request URL, check if an image has been previously downloaded, and if not, downloads it and updates the download cache file. The download location is specified on the local machine and is configurable.